[Security Solution] fix missing privilege in new flyout not in the correct location
