+++ x-pack/solutions/security/plugins/security_solution/public/flyout/document_details/shared/context.tsx
@@ -13,7 +13,7 @@ import { DEFAULT_ALERTS_INDEX } from '../../../../common/constants';
 import { useEventDetails } from './hooks/use_event_details';
 import { FlyoutError } from '../../shared/components/flyout_error';
 import { FlyoutLoading } from '../../shared/components/flyout_loading';
-import { FlyoutMissingAlertsPrivilege } from '../../shared/components/flyout_missing_alerts_privilege';
+import { FlyoutMissingAlertsPrivilege } from '../../../flyout_v2/document/components/flyout_missing_alerts_privilege';
 import type { SearchHit } from '../../../../common/search_strategy';
 import { useBasicDataFromDetailsData } from './hooks/use_basic_data_from_details_data';
 import type { DocumentDetailsProps } from './types';
