+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/document/components/test_ids.ts
@@ -7,6 +7,8 @@
 
 import { PREFIX } from '../../../flyout/shared/test_ids';
 
+export const FLYOUT_MISSING_ALERTS_PRIVILEGE_TEST_ID = `${PREFIX}MissingAlertsPrivilege` as const;
+
 /* Header */
 
 export const HEADER_TITLE_TEST_ID = `${PREFIX}AlertTitle` as const;
