+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/document/index.test.tsx
@@ -1,0 +1,58 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import React from 'react';
+import { render } from '@testing-library/react';
+import type { DataTableRecord } from '@kbn/discover-utils';
+import { useAlertsPrivileges } from '../../detections/containers/detection_engine/alerts/use_alerts_privileges';
+import { DocumentFlyout } from '.';
+import { TestProviders } from '../../common/mock';
+
+jest.mock('../../detections/containers/detection_engine/alerts/use_alerts_privileges');
+jest.mock('./header', () => ({ Header: () => <div data-test-subj="mock-header" /> }));
+jest.mock('./tabs/overview_tab', () => ({
+  OverviewTab: () => <div data-test-subj="mock-overview-tab" />,
+}));
+
+const createAlertHit = (): DataTableRecord =>
+  ({
+    id: '1',
+    raw: {},
+    flattened: { 'event.kind': 'signal' },
+    isAnchor: false,
+  } as DataTableRecord);
+
+describe('<DocumentFlyout />', () => {
+  beforeEach(() => {
+    jest.clearAllMocks();
+  });
+
+  it('renders FlyoutMissingAlertsPrivilege when document is an alert and user lacks alerts read privilege', () => {
+    (useAlertsPrivileges as jest.Mock).mockReturnValue({ hasAlertsRead: false, loading: false });
+
+    const { getByTestId } = render(
+      <TestProviders>
+        <DocumentFlyout hit={createAlertHit()} renderCellActions={jest.fn()} />
+      </TestProviders>
+    );
+
+    expect(getByTestId('noPrivilegesPage')).toBeInTheDocument();
+  });
+
+  it('renders loading while alerts privileges are loading for an alert', () => {
+    (useAlertsPrivileges as jest.Mock).mockReturnValue({ hasAlertsRead: false, loading: true });
+
+    const { getByTestId, queryByTestId } = render(
+      <TestProviders>
+        <DocumentFlyout hit={createAlertHit()} renderCellActions={jest.fn()} />
+      </TestProviders>
+    );
+
+    expect(getByTestId('document-overview-loading')).toBeInTheDocument();
+    expect(queryByTestId('noPrivilegesPage')).not.toBeInTheDocument();
+  });
+});
